refactor(ui): address PR review findings
- Extract format_relative_time to shared views/helpers.rs (was duplicated 3x)
- Fix misleading comments in detail_view (status dot, hover visibility)
- Remove redundant section label comments in detail_view
- Extract note truncation magic number to MAX_NOTE_LENGTH constant
- Replace renaming_tab 3-tuple with named RenamingTab struct
