+++ crates/kild-ui/src/views/mod.rs
@@ -10,6 +10,7 @@
 //! - `confirm_dialog` - Modal dialog for confirming destructive actions
 //! - `add_project_dialog` - Modal dialog for adding new projects
 //! - `sidebar` - Fixed left sidebar for project navigation
+//! - `helpers` - Shared view utilities (time formatting, etc.)
 
 pub mod add_project_dialog;
 pub mod confirm_dialog;
@@ -17,6 +18,7 @@ pub mod create_dialog;
 pub mod dashboard_view;
 pub mod detail_panel;
 pub mod detail_view;
+pub mod helpers;
 pub mod kild_list;
 pub mod main_view;
 pub mod project_rail;
